move utils to folder inside project to avoid strange folder structure
