+++ packages/dialtone-vue2/components/modal/modal.vue
@@ -123,13 +123,12 @@ import {
   MODAL_KIND_MODIFIERS,
   MODAL_SIZE_MODIFIERS,
 } from './modal_constants';
-import { getUniqueString } from '@/common/utils';
+import { getUniqueString, disableRootScrolling, enableRootScrolling } from '@/common/utils';
 import { DtLazyShow } from '@/components/lazy_show';
 import { EVENT_KEYNAMES } from '@/common/constants';
 import SrOnlyCloseButtonMixin from '@/common/mixins/sr_only_close_button';
 import SrOnlyCloseButton from '@/common/sr_only_close_button.vue';
 import { NOTICE_KINDS } from '@/components/notice';
-import { disableRootScrolling, enableRootScrolling } from '@/../../common/utils';
 
 /**
  * Modals focus the user’s attention exclusively on one task or piece of information
