+++ packages/dialtone-vue3/components/popover/popover.vue
@@ -135,7 +135,7 @@ import {
   POPOVER_ROLES,
   POPOVER_STICKY_VALUES,
 } from './popover_constants';
-import { getUniqueString, hasSlotContent, isOutOfViewPort, warnIfUnmounted } from '@/common/utils';
+import { getUniqueString, hasSlotContent, isOutOfViewPort, warnIfUnmounted, disableRootScrolling, enableRootScrolling } from '@/common/utils';
 import { DtLazyShow } from '@/components/lazy_show';
 import ModalMixin from '@/common/mixins/modal';
 import { createTippyPopover, getPopperOptions } from './tippy_utils';
@@ -143,7 +143,6 @@ import PopoverHeaderFooter from './popover_header_footer.vue';
 import SrOnlyCloseButtonMixin from '@/common/mixins/sr_only_close_button';
 import SrOnlyCloseButton from '@/common/sr_only_close_button.vue';
 import { TOOLTIP_DELAY_MS } from '@/components/tooltip/index.js';
-import { disableRootScrolling, enableRootScrolling } from '@/../../common/utils';
 
 /**
  * A Popover displays a content overlay when its anchor element is activated.
